Unpick circular dependencies between export/items routes by moving lookups to mongo module
